online_portfolio/backend/requirements.txt
appdirs==1.4.3
arrow==0.15.4
asgiref==3.2.3
attrs==19.3.0
beautifulsoup4==4.8.1
black==19.10b0
boltons==19.3.0
certifi==2019.9.11
chardet==3.0.4
Click==7.0
Django==3.0.2
django-cors-headers==3.2.1
django-crispy-forms==1.8.1
django-filter==2.2.0
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
entrypoints==0.3
flake8==3.7.9
idna==2.8
lxml==4.4.1
mccabe==0.6.1
numpy==1.17.4
pathspec==0.6.0
Pillow==7.0.0
pipenv==2018.11.26
pycodestyle==2.5.0
pyflakes==2.1.1
PyJWT==1.7.1
python-dateutil==2.8.1
pytz==2019.3
regex==2019.11.1
requests==2.22.0
ruamel.yaml==0.16.5
ruamel.yaml.clib==0.2.0
six==1.13.0
soupsieve==1.9.5
sqlparse==0.3.0
toml==0.10.0
typed-ast==1.4.0
urllib3==1.25.6
virtualenv==16.7.9
virtualenv-clone==0.5.3
